+++ backend/src/test/java/com/woowacourse/ternoko/acceptance/CommentAcceptanceTest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.ternoko.comment.acceptance;
+package com.woowacourse.ternoko.acceptance;
 
 import static com.woowacourse.ternoko.support.fixture.MemberFixture.COACH1;
 import static com.woowacourse.ternoko.support.fixture.MemberFixture.CREW1;
@@ -6,13 +6,12 @@ import static com.woowacourse.ternoko.support.fixture.MemberFixture.CREW2;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import com.woowacourse.ternoko.acceptance.AcceptanceSupporter;
-import com.woowacourse.ternoko.core.presentation.request.CommentRequest;
 import com.woowacourse.ternoko.core.application.response.CommentResponse;
 import com.woowacourse.ternoko.core.application.response.CommentsResponse;
-import com.woowacourse.ternoko.core.domain.member.MemberType;
-import com.woowacourse.ternoko.core.domain.interview.InterviewStatusType;
 import com.woowacourse.ternoko.core.application.response.InterviewResponse;
+import com.woowacourse.ternoko.core.domain.interview.InterviewStatusType;
+import com.woowacourse.ternoko.core.domain.member.MemberType;
+import com.woowacourse.ternoko.core.presentation.request.CommentRequest;
 import io.restassured.http.Header;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
