+++ backend/src/test/java/com/woowacourse/ternoko/api/CommentControllerTest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.ternoko.comment.presentation;
+package com.woowacourse.ternoko.api;
 
 import static com.woowacourse.ternoko.auth.application.AuthorizationExtractor.AUTHORIZATION;
 import static com.woowacourse.ternoko.auth.application.AuthorizationExtractor.BEARER_TYPE;
@@ -11,11 +11,11 @@ import static org.mockito.Mockito.doNothing;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.woowacourse.ternoko.core.domain.comment.Comment;
-import com.woowacourse.ternoko.core.presentation.request.CommentRequest;
 import com.woowacourse.ternoko.core.application.response.CommentsResponse;
-import com.woowacourse.ternoko.core.presentation.CommentController;
+import com.woowacourse.ternoko.core.domain.comment.Comment;
 import com.woowacourse.ternoko.core.domain.member.MemberType;
+import com.woowacourse.ternoko.core.presentation.CommentController;
+import com.woowacourse.ternoko.core.presentation.request.CommentRequest;
 import com.woowacourse.ternoko.support.utils.WebMVCTest;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
