+++ backend/src/test/java/com/woowacourse/ternoko/api/restdocs/RestDocsTestSupporter.java
@@ -6,6 +6,7 @@ import static org.mockito.BDDMockito.given;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.woowacourse.ternoko.auth.application.AuthService;
 import com.woowacourse.ternoko.auth.application.JwtProvider;
+import com.woowacourse.ternoko.auth.config.AuthenticationPrincipalConfig;
 import com.woowacourse.ternoko.auth.presentation.LoginMemberVerifier;
 import com.woowacourse.ternoko.auth.presentation.MemberTypeCache;
 import com.woowacourse.ternoko.common.log.CustomLoggingFilter;
@@ -14,10 +15,9 @@ import com.woowacourse.ternoko.config.WebConfig;
 import com.woowacourse.ternoko.core.application.CoachService;
 import com.woowacourse.ternoko.core.application.CommentService;
 import com.woowacourse.ternoko.core.application.CrewService;
-import com.woowacourse.ternoko.support.application.EmailService;
 import com.woowacourse.ternoko.core.application.InterviewService;
 import com.woowacourse.ternoko.core.application.MemberService;
-import com.woowacourse.ternoko.auth.config.AuthenticationPrincipalConfig;
+import com.woowacourse.ternoko.support.application.EmailService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
