+++ backend/src/test/java/com/woowacourse/ternoko/domain/comment/CommentTest.java
@@ -1,4 +1,4 @@
-package com.woowacourse.ternoko.comment.domain;
+package com.woowacourse.ternoko.domain.comment;
 
 //TODO: CommentServiceTest 에서 도메인 관련 검증로직 옮겨오기
 public class CommentTest {
