+++ backend/src/test/java/com/woowacourse/ternoko/service/InterviewServiceTest.java
@@ -1,12 +1,12 @@
-package com.woowacourse.ternoko.interview.application;
+package com.woowacourse.ternoko.service;
 
-import static com.woowacourse.ternoko.core.domain.availabledatetime.AvailableDateTimeStatus.OPEN;
 import static com.woowacourse.ternoko.common.exception.ExceptionType.INTERVIEW_NOT_FOUND;
 import static com.woowacourse.ternoko.common.exception.ExceptionType.INVALID_AVAILABLE_DATE_TIME;
 import static com.woowacourse.ternoko.common.exception.ExceptionType.INVALID_INTERVIEW_COACH_ID;
 import static com.woowacourse.ternoko.common.exception.ExceptionType.INVALID_INTERVIEW_CREW_ID;
 import static com.woowacourse.ternoko.common.exception.ExceptionType.INVALID_INTERVIEW_DATE;
 import static com.woowacourse.ternoko.common.exception.ExceptionType.INVALID_INTERVIEW_DUPLICATE_DATE_TIME;
+import static com.woowacourse.ternoko.core.domain.availabledatetime.AvailableDateTimeStatus.OPEN;
 import static com.woowacourse.ternoko.support.fixture.CoachAvailableTimeFixture.FIRST_TIME;
 import static com.woowacourse.ternoko.support.fixture.CoachAvailableTimeFixture.MONTH_REQUEST;
 import static com.woowacourse.ternoko.support.fixture.CoachAvailableTimeFixture.NOW;
@@ -34,22 +34,22 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import com.woowacourse.ternoko.core.domain.availabledatetime.AvailableDateTime;
-import com.woowacourse.ternoko.core.presentation.request.AvailableDateTimeRequest;
-import com.woowacourse.ternoko.core.presentation.request.AvailableDateTimeSummaryRequest;
-import com.woowacourse.ternoko.core.domain.availabledatetime.AvailableDateTimeRepository;
-import com.woowacourse.ternoko.core.application.InterviewService;
-import com.woowacourse.ternoko.core.presentation.request.CalendarRequest;
-import com.woowacourse.ternoko.core.domain.interview.InterviewStatusType;
-import com.woowacourse.ternoko.core.application.response.FormItemResponse;
-import com.woowacourse.ternoko.core.presentation.request.InterviewRequest;
-import com.woowacourse.ternoko.core.application.response.InterviewResponse;
-import com.woowacourse.ternoko.core.application.response.ScheduleResponse;
 import com.woowacourse.ternoko.common.exception.InterviewNotFoundException;
 import com.woowacourse.ternoko.common.exception.InvalidInterviewCoachIdException;
 import com.woowacourse.ternoko.common.exception.InvalidInterviewCrewIdException;
 import com.woowacourse.ternoko.common.exception.InvalidInterviewDateException;
 import com.woowacourse.ternoko.core.application.CoachService;
+import com.woowacourse.ternoko.core.application.InterviewService;
+import com.woowacourse.ternoko.core.application.response.FormItemResponse;
+import com.woowacourse.ternoko.core.application.response.InterviewResponse;
+import com.woowacourse.ternoko.core.application.response.ScheduleResponse;
+import com.woowacourse.ternoko.core.domain.availabledatetime.AvailableDateTime;
+import com.woowacourse.ternoko.core.domain.availabledatetime.AvailableDateTimeRepository;
+import com.woowacourse.ternoko.core.domain.interview.InterviewStatusType;
+import com.woowacourse.ternoko.core.presentation.request.AvailableDateTimeRequest;
+import com.woowacourse.ternoko.core.presentation.request.AvailableDateTimeSummaryRequest;
+import com.woowacourse.ternoko.core.presentation.request.CalendarRequest;
+import com.woowacourse.ternoko.core.presentation.request.InterviewRequest;
 import com.woowacourse.ternoko.support.utils.DatabaseSupporter;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
